Move webgl files into a folder
