+++ src/assignment7/webgl/Light.ts
